move getWordAt to utils
