+++ src/mentionPlugin/utils/getWordAt.js
